changed signed/unsigned warning

+++ robotic-cleaner-tester/robotic-cleaner-tester/Battery.cpp
@@ -45,7 +45,7 @@ namespace ns_robotic_cleaner_simulator
 	void Battery::Recharge()
 	{
 		_battery_level = _battery_level + _recharge_rate;
-		if(_battery_level > _battery_capacity)
+		if(_battery_level > (int)_battery_capacity)
 			_battery_level = _battery_capacity;
 	}
